+++ docs/docs/feliz-docs/Examples/Feliz/ReactComponentTransform/RecordType.fs
@@ -1,0 +1,16 @@
+module Example.ReactComponentTransform.RecordType
+
+type Props = {
+    text: string
+    count: int
+}
+
+open Feliz
+open Feliz.JSX
+
+[<ReactComponent>]
+let Component (recordType: Props) =
+    Html.div [
+        Html.text recordType.text
+        Html.text recordType.count
+    ]