+++ playground/src/Main.fs
@@ -4,4 +4,8 @@ open Feliz
 open Browser.Dom
 
 let private root = ReactDOM.createRoot(document.getElementById "root")
-root.render(App.App())
+root.render(
+    React.StrictMode [
+        App.App()
+    ]
+)