+++ src/Feliz.Template/Content/vite.config.js
@@ -14,4 +14,9 @@ export default defineConfig({
     environment: 'jsdom',
     setupFiles: ['./vitest-setup.ts'],
   },
+  server: {
+      watch: {
+          ignored: [ "**/*.fs" ]
+      },
+  }
 })