+++ src/Feliz.UseElmish/CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 4.0.0-rc.2 - 2025-10-02
+
+### ✨ Added 
+
+- Additional overloads for subscriptions (by @auslavs)
+
 ## 4.0.0-rc.1 - 2025-10-02
 
 ### Changed